[rename] 파일명 수정 및 안쓰는 파일 삭제
