+++ apps/frontend/src/components/LoginModal.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { ModalStep } from '@/constants/modalConstants';
+import { ModalStep } from '@/constants/ModalConstants';
 
 interface LoginProps {
   setModalStep: (step: ModalStep) => void;
