+++ apps/frontend/src/types/index.ts
@@ -1,4 +1,4 @@
-export * from '@/types/alarm';
-export * from '@/types/rank';
-export * from '@/types/transaction';
-export * from '@/types/crop';
+export * from '@/types/Alarm';
+export * from '@/types/Rank';
+export * from '@/types/Transaction';
+export * from '@/types/Crop';
